Deleted SelectionFragment.java (Fragment) and replaced it with Product.java (Activity).
Added functionality to the buttons in Product.java. Clicking either
button now changes the image to a random image in the res/drawable
directory. Also added functionality to change orientation without it
changing the image. This was done by adding the line
android:configChanges="orientation | screenSize" to
AndroidManifest.xml. Added the functionality of pressing the back
button to close the app.